Add files via upload
This is a Bash script for automating the process of changing passwords for expired accounts. I have saved this script on my Linux server with the name 'pass_changer.' When an account expires, I can simply run this script and enter the username of the expired password. The result will provide the username along with the new password for the user.